fix(plugin): export only default from entry module (OpenCode 1.18.4)
OpenCode 1.18.4's loader invokes EVERY function-valued export of a plugin
entry module as a plugin factory (iterates Object.values(entryModule) and
calls each with PluginInput). index.ts also exported scrubSecrets and
callMorphApply, which OpenCode invoked as bogus factories. They return a
string / result object rather than throwing, so this produced no load error
— but it silently registered garbage plugins alongside the real one.

Fix (matches the toolbox plugin fixes): entry module exports exactly one
thing — default. All implementation + named test exports move to impl.ts via
git mv (no logic change); index.ts becomes a thin shim importing impl.ts and
re-exporting default. Repoint src/execute.ts + index.test.ts imports to
impl.js; add impl.ts to package.json files.

137/137 tests pass; typecheck clean; entry verified to export exactly
[default].
